requirements.txt
torch==2.0.1
pyyaml==6.0
torchaudio==2.0.2
numpy==1.24.3
pandas==2.0.1
scikit-learn==1.2.2
pytest~=7.3.1
oyaml==1.0
mlflow==2.3.1
clearml==1.10.4
dvc==2.56.0
librosa~=0.10.0.post2